QAmqpQueuePrivate: Convert to state machines.
We have split this into three state machines:
- Queue state machine, just manages declaration and removal of a queue.
- Consumer state machine focusses on the consumer state.
- Binding state machine looks after binding and unbinding.

The Declare-OK, Consume-OK, Bind-OK, Unbind-OK signals all trigger the
Binding state machine to check bindings and issue Bind/Unbind requests
as needed.  On the TODO list is to hook the 'declared' signal from
exchanges.

The consumer state machine is triggered by Declare-OK and Consume-OK,
and it supports delayed consuming, so that a program can request
consumation and it'll take place once the queue is declared.