Update qamqpframe.cpp
I have a problem with method waitForBytesWritten.
Sometimes my application is locked when is called this method (timeout expired).
Although I try to modify writeTimeout (i.e. -1; 5 or 30), when timeout expires, rabbitMQ server disconnects me.

But if I try to delete it, it works!
My question is:
Why is this method necessary? What is its purpose?

Thank you!